main/requirements.txt
uvicorn
fastapi
mysqlclient
sqlalchemy
psycopg2
SQLAlchemy-Utils
alembic
python-dotenv
pydantic[email]